added suffix to NumberSetting, so you can add "ms" to a delay setting for example

+++ odinmain/src/main/kotlin/me/odinmain/ui/clickgui/animations/impl/LinearAnimation.kt
@@ -1,17 +1,12 @@
 package me.odinmain.ui.clickgui.animations.impl
 
 import me.odinmain.ui.clickgui.animations.Animation
-import me.odinmain.utils.minus
-import me.odinmain.utils.plus
-import me.odinmain.utils.times
 
 class LinearAnimation<E>(duration: Long): Animation<E>(duration) where E : Number, E: Comparable<E> {
 
-    // x + 10f (x + 220f - x + 10f) * sat
-
     override fun get(start: E, end: E, reverse: Boolean): E {
-        val startVal = if (reverse) end else start
-        val endVal = if (reverse) start else end
+        val startVal = if (reverse) end.toFloat() else start.toFloat()
+        val endVal = if (reverse) start.toFloat()  else end.toFloat()
 
         if (!isAnimating()) return if (reverse) start else end
         return (startVal + (endVal - startVal) * (getPercent() / 100f)) as E
